Refactor Notification API. Move code to it's own file.
